Changes for unlink if condition

+++ app/Http/Services/AboutUs/StateDisasterManagementAuthorityServices.php
@@ -62,7 +62,9 @@ class StateDisasterManagementAuthorityServices
             $path = Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_ADD');
             if ($request->hasFile('english_image')) {
                 if ($return_data['english_image']) {
-                    unlink(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['english_image']));
+                    if (file_exists(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['english_image']))) {
+                        unlink(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['english_image']));
+                    }
 
                 }
                 $englishImageName = $return_data['last_insert_id'] . '_english.' . $request->english_image->extension();
@@ -74,7 +76,9 @@ class StateDisasterManagementAuthorityServices
     
             if ($request->hasFile('marathi_image')) {
                 if ($return_data['marathi_image']) {
-                    unlink(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['marathi_image']));
+                    if (file_exists(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['marathi_image']))) {
+                        unlink(storage_path(Config::get('DocumentConstant.ABOUT_US_STATE_DISASTER_MGTAUTHORITY_DELETE') . $return_data['marathi_image']));
+                    }
                 }
                 $marathiImageName = $return_data['last_insert_id'] . '_marathi.' . $request->marathi_image->extension();
                 uploadImage($request, 'marathi_image', $path, $marathiImageName);
